ipc 리펙터링 및 실시간 채팅방 정보 업데이트

+++ backend/crates/talk-loco-client/src/talk/stream/command.rs
@@ -10,7 +10,7 @@ pub struct Kickout {
     /// * Change server = 2
     /// * Login another = 0
     /// * Account deleted = 1
-    pub reason: i16,
+    pub reason: i32,
 }
 
 /// Message sent from chatroom
